test: move tests of test/test_stat.rs to test/sys/test_stat.rs
